添加邀请码有效查询

+++ src/main/java/com/muma/entity/UserDetail.java
@@ -133,7 +133,7 @@ public class UserDetail extends BaseEntity {
     return parentId;
   }
 
-  public void setParentId(Long parentId) {
+  public void setParentId(Integer parentId) {
     this.parentId = parentId;
   }
